requirements.txt
google-cloud-translate==1.5.0
